rename: api 디렉터리 내 파일들 관심사에 맞게 분리
