+++ src/api/utils/fetchClient.ts
@@ -1,6 +1,6 @@
 import { API_URL, USER_ID, USER_PASSWORD } from '@/api/config';
 
-import { generateBasicToken } from '@/utils/auth';
+import generateBasicToken from '@/api/utils/generateBasicToken';
 
 const fetchFromAPI = async (endpoint: string, options: RequestInit) => {
   const response = await fetch(`${API_URL}${endpoint}`, options);
@@ -9,7 +9,7 @@ const fetchFromAPI = async (endpoint: string, options: RequestInit) => {
 
 export const fetchWithAuth = async (
   endpoint: string,
-  options: RequestInit = {}
+  options: RequestInit = {},
 ) => {
   const token = generateBasicToken(USER_ID, USER_PASSWORD);
   const headersWithAuth = {
@@ -21,7 +21,7 @@ export const fetchWithAuth = async (
 
 export const fetchWithoutAuth = async (
   endpoint: string,
-  options: RequestInit = {}
+  options: RequestInit = {},
 ) => {
   return fetchFromAPI(endpoint, options);
 };
