+++ src/hooks/useProductSelector.ts
@@ -1,8 +1,8 @@
-import { deleteCartItem, postCartItem } from '@/api/cartItem';
 import { useEffect, useState } from 'react';
 
 import useCartItemFinder from '@/hooks/useCartItemFinder';
 import useCartListContext from '@/hooks/useCartListContext';
+import { deleteCartItem, postCartItem } from '@/api/cartItem';
 
 const useProductSelector = (productId: number) => {
   const [loading, setLoading] = useState(false);
